requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
python-dotenv==1.2.1
requests==2.32.5
openai==2.15.0